// src/app/shared/model/attr-content.model.ts
export class AttrContent {
    CustAttrContentId: number;
    CustId: number;
    RefAttrId: number;
    AttrGroup: string;
    AttrCode: string;
    AttrName: string;  
    AttrInputType: string;
    AttrValue: string;
    MasterCode: string;
    Descr : string;
    PatternValue : string;
    IsMandatory: boolean; 
    DefaultValue: string; 
    RowVersion: string; 
    constructor() { 
     }
}  