// src/app/shared/model/ref-attr.model.ts
export class RefAttr {
    RefAttrId: number;
    AttrCode: string;
    AttrName: string;
    IsActive: boolean;
    IsSystem: boolean;
    AttrLength: number;
    AttrTypeCode: string;
    AttrInputType: string;
    AttrValue: string;
    AttrGroup: string;
    RsvField1: string;
    RsvField2: string;
    RsvField3: string;
    RsvField4: string;
    RsvField5: string;
    PatternCode: string;
    PatternValue: string;
    IsMandatory: boolean;
    DefaultValue: string; 
    constructor() { }
}